Automatically tune Spark based on instance type
Uses the formulae at https://aws.amazon.com/blogs/big-data/best-practices-for-successfully-managing-memory-for-apache-spark-applications-on-amazon-emr/
to automatically tune Spark's memory and parallelism settings based on the number
and type of core nodes.

For this to work, though, we need to know which instance type EMR is going to
launch. So this also uses a single instance fleet configuration for the core nodes.