DW-5094: Hive Metastore CloudWatch Logs and Insights (#392)
* DW-5094: Hive Metastore CloudWatch Logs and Insights

This adds support for shipping MySQL logs from the Hive Metastore
cluster to CloudWatch. It also enables Peformance Insights so we
can get visibility of potential performance issues.

Finally, it adds supporting infra to enable Enhanced Monitoring
should the above tools hint that it would provide value. As we
don't yet know whether it's valuable, or at what frequency we should
collect metrics, it's simply disabled for now, but easily enabled
on a per-env basis.

* DW-5094: Move RDS EM role to common-infra