DW-5904: Enable performance insights in QA (#401)
This ensures we have at least one other env with the
same config as prod.

Also, export the error logs to CloudWatch. The docs say that they're
exported by default, but if you choose to export some non-default logs
then that overrides *that* default so they end up *not* being exported.